We cannot simply use Byte.toInt().
Why? Because Byte values can be negative!

So, if we simply use Byte.toInt() just to convert it to Int like C# where
`byte` is **unsigned**, we get unexpected negative int values.

The idea that Byte should be *signed* by default is really a stupid idea,
but it's been since Java language was designed, so it is irrecoverable...